requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
Booktype==1.5
certifi==2024.8.30
cffi==1.17.1
Django==5.1.1
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
minio==7.2.12
packaging==24.2
psycopg2==2.9.10
pycparser==2.22
pycryptodome==3.21.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
setuptools==75.8.0
simplejson==3.19.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3

segno~=1.6.1